pyproject.toml
[tool.poetry]
name = "codes-ai"
version = "0.1.0"
description = ""
authors = ["Oktapian <oktapian@jkt1.ebdesk.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
module-typica = {version='0.1.2', source='ebdesk'}
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
pymongo = "^4.6.0"
redis = "^5.0.1"
pika = "^1.3.2"
loguru = "^0.7.2"
pyjwt = "^2.8.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
xmltodict = "^0.13.0"
requests = "^2.31.0"
mysql-connector-python = "^8.2.0"
python-dateutil = "^2.8.2"
sastrawi = "^1.0.1"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "ebdesk"
url = "http://192.168.20.127:6060/simple"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
